refactor: more package organization
